Rename pyswervedrive modules to be non-repetitive
